+++ src/info.py
@@ -1,4 +1,4 @@
 class Info:
     AppName = 'S-Parameter Viewer'
-    AppVersionStr = "0.13b1" # see https://peps.python.org/pep-0440/
-    AppDateStr = "2023-07-27"
+    AppVersionStr = "0.14b1" # see https://peps.python.org/pep-0440/
+    AppDateStr = "2023-09-08"